Remove WPF and add AvaloniaUI
