+++ AvaloniaUIClient/App.axaml
@@ -1,0 +1,11 @@
+<Application
+    x:Class="AvaloniaUIClient.App"
+    xmlns="https://github.com/avaloniaui"
+    xmlns:local="using:AvaloniaUIClient"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <Application.Styles>
+        <FluentTheme Mode="Dark" />
+    </Application.Styles>
+
+</Application>
