+++ AvaloniaUIClient/Views/MainWindow.axaml
@@ -1,0 +1,19 @@
+<Window
+    Icon="/Assets/avalonia-logo.ico"
+    Title="AvaloniaUIClient"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    mc:Ignorable="d"
+    x:Class="AvaloniaUIClient.Views.MainWindow"
+    xmlns="https://github.com/avaloniaui"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+
+    <TextBlock
+        HorizontalAlignment="Center"
+        Text="{Binding Greeting}"
+        VerticalAlignment="Center" />
+
+</Window>
