--- WPFClient/App.xaml
@@ -1,24 +1,0 @@
-﻿<Application
-    x:Class="MVVMBase.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:local="clr-namespace:MVVMBase"
-    xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-
-                <materialDesign:BundledTheme
-                    BaseTheme="Dark"
-                    PrimaryColor="LightBlue"
-                    SecondaryColor="LightBlue" />
-
-                <ResourceDictionary Source="pack://application:,,,/MaterialDesignThemes.Wpf;component/Themes/MaterialDesignTheme.Defaults.xaml" />
-
-                <ResourceDictionary Source="Resources/Styles/LogEventStyle.xaml" />
-
-            </ResourceDictionary.MergedDictionaries>
-
-        </ResourceDictionary>
-    </Application.Resources>
-</Application>
