--- WPFClient/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace MVVMBase;
-
-public class Program
-{
-    [STAThread]
-    public static void Main()
-    {
-        var app = new App();
-
-        app.InitializeComponent();
-
-        app.Run();
-    }
-}
