--- WPFClient/Resources/Styles/LogEventStyle.xaml
@@ -1,10 +1,0 @@
-﻿<ResourceDictionary
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:events="clr-namespace:Serilog.Events;assembly=Serilog"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-
-    <DataTemplate DataType="{x:Type events:LogEvent}">
-        <TextBlock Text="{Binding Exception.Message}" />
-    </DataTemplate>
-
-</ResourceDictionary>
