--- WPFClient/Views/MainWindow.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows;
-
-namespace MVVMBase.Views
-{
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-    }
-}
